backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
python-dotenv==1.1.1
Pillow==11.3.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
requests-oauthlib==2.0.0
